tweak: removed v1 module, and moved deprecated files to `contracts/deprecated`
