+++ packages/foundry/contracts/deprecated/FYTokenFactoryMock.sol
