+++ packages/foundry/contracts/deprecated/IPoolFactory.sol
@@ -1,0 +1,17 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+
+interface IPoolFactory {
+  event PoolCreated(address indexed base, address indexed fyToken, address pool);
+
+  function POOL_BYTECODE_HASH() external pure returns (bytes32);
+  function calculatePoolAddress(address base, address fyToken) external view returns (address);
+  function getPool(address base, address fyToken) external view returns (address);
+  function createPool(address base, address fyToken) external returns (address);
+  function nextBase() external view returns (address);
+  function nextFYToken() external view returns (address);
+  function ts() external view returns (int128);
+  function g1() external view returns (int128);
+  function g2() external view returns (int128);
+}
