+++ packages/foundry/contracts/deprecated/Wand.sol
@@ -9,9 +9,9 @@ import "@yield-protocol/vault-interfaces/src/IJoin.sol";
 import "@yield-protocol/vault-interfaces/src/IFYTokenFactory.sol";
 import "@yield-protocol/vault-interfaces/src/IFYToken.sol";
 import "@yield-protocol/vault-interfaces/src/DataTypes.sol";
-import "@yield-protocol/yieldspace-interfaces/IPoolFactory.sol";
+import "./IPoolFactory.sol";
 import "@yield-protocol/utils-v2/contracts/access/AccessControl.sol";
-import "./constants/Constants.sol";
+import "../constants/Constants.sol";
 
 interface IWitchGov {
     function limits(bytes6) external view returns(uint96, uint24, uint8, uint128);
