+++ packages/foundry/contracts/mocks/PoolMock.sol
@@ -4,7 +4,7 @@ import "@yield-protocol/utils-v2/contracts/access/Ownable.sol";
 import "@yield-protocol/utils-v2/contracts/token/IERC20.sol";
 import "@yield-protocol/utils-v2/contracts/token/ERC20.sol";
 import "@yield-protocol/vault-interfaces/src/IFYToken.sol";
-import "@yield-protocol/yieldspace-interfaces/IPoolFactory.sol";
+import "../deprecated/IPoolFactory.sol";
 import "./ERC20Mock.sol";
