+++ packages/hardhat/package.json
@@ -6,13 +6,14 @@
   "files": [
     "contracts/*.sol",
     "contracts/**/*.sol",
-    "!contracts/mocks/*.sol"
+    "!contracts/mocks/**/*.sol",
+    "!contracts/deprecated/**/*.sol"
   ],
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "scripts": {
     "hardhat:build": "rm -rf artifacts typechain && npx ts-node ./src/importer.ts && hardhat compile",
-    "hardhat:test": "yarn build && hardhat test --show-stack-traces",
+    "hardhat:test": "hardhat compile && hardhat test --show-stack-traces",
     "hardhat:test:05": "hardhat test test/01* test/02* test/03* test/05*",
     "hardhat:test:09": "hardhat test test/06* test/07* test/08* test/09*",
     "hardhat:test:10": "hardhat test test/10*",
