+++ packages/hardhat/test/shared/contango_fixtures.ts
@@ -6,16 +6,16 @@ import CauldronArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts
 import WitchArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/other/contango/ContangoWitch.sol/ContangoWitch.json'
 import DAIMockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/DAIMock.sol/DAIMock.json'
 import ERC20MockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/ERC20Mock.sol/ERC20Mock.json'
-import JoinFactoryArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/JoinFactoryMock.sol/JoinFactoryMock.json'
+import JoinFactoryArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/deprecated/JoinFactoryMock.sol/JoinFactoryMock.json'
 import ChainlinkAggregatorV3MockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/oracles/chainlink/ChainlinkAggregatorV3Mock.sol/ChainlinkAggregatorV3Mock.json'
 import CTokenChiMockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/oracles/compound/CTokenChiMock.sol/CTokenChiMock.json'
 import CTokenRateMockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/oracles/compound/CTokenRateMock.sol/CTokenRateMock.json'
-import PoolFactoryMockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/PoolFactoryMock.sol/PoolFactoryMock.json'
+import PoolFactoryMockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/deprecated/PoolFactoryMock.sol/PoolFactoryMock.json'
 import USDCMockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/USDCMock.sol/USDCMock.json'
 import WETH9MockArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/mocks/WETH9Mock.sol/WETH9Mock.json'
 import ChainlinkMultiOracleArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/oracles/chainlink/ChainlinkMultiOracle.sol/ChainlinkMultiOracle.json'
 import CompoundMultiOracleArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/oracles/compound/CompoundMultiOracle.sol/CompoundMultiOracle.json'
-import WandArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/Wand.sol/Wand.json'
+import WandArtifact from '../../artifacts/@yield-protocol/vault-v2/contracts/deprecated/Wand.sol/Wand.json'
 import { CHI, DAI, ETH, RATE, USDC } from '../../src/constants'
 import { Cauldron, ContangoLadle, ContangoWitch } from '../../typechain'
 import { ChainlinkMultiOracle } from '../../typechain/ChainlinkMultiOracle'
